+++ View/view.py
@@ -312,7 +312,7 @@ class View(QObject):
     def color_id_to_color_list(self, id_list):
         # try:
         if id_list.shape[-1] == 1:
-            id_list = id_list.reshape(-1)
+            id_list = id_list.reshape(-1).astype(np.uint8)
             color_dim = 3
             rgb_color_map = {}
             for key, value in self.color_map.items():
@@ -331,7 +331,7 @@ class View(QObject):
                 color_dim = len(rgb_color_map[key])
             ret_color = np.zeros((len(id_list), color_dim))
             for key, value in rgb_color_map.items():
-                mask = (id_list[:, 0] == int(key))
+                mask = (id_list[:, 0].astype(np.uint8) == int(key))
                 ret_color[mask] = value
                 ret_color[mask, -1] = np.clip(id_list[mask, -1], 0.0, 1.0)
             return ret_color, True
